Test: 테스트 코드 추가 (#190)
* Test: 로그인 계정 정보 , 투자 성향 테스트 단위 + 통합 테스트

* Test: 백오피스 예약건수+리스트 불러오기 통합 + 단위 테스트, PB 회원가입
테스트시 s3에 사진 업로드 되고 삭제하는 기능 추가

* Fix: test 오류 해결

---------

+++ src/main/java/kr/co/moneybridge/model/reservation/ReservationRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 
 public interface ReservationRepository extends JpaRepository<Reservation, Long> {
     @Query(value = "select count(r) from Reservation r where r.process = :process")
-    Integer countByProcess(@Param("process") ReservationProcess process);
+    Long countByProcess(@Param("process") ReservationProcess process);
 
     @Modifying
     @Query("delete from Reservation r where r.user.id = :userId")
